refactor: implement api package in web app
