+++ apps/web/src/app/_trpc/client.ts
@@ -1,7 +1,13 @@
-import { type AppRouter } from "@/server/router";
+// import { type AppRouter } from "@/server/router";
+
+// import { createTRPCReact } from "@trpc/react-query";
+
+// export const trpc = createTRPCReact<AppRouter>({
+//   abortOnUnmount: true,
+// });
+
+import { type AppRouter } from "@taskaider/api";
 
 import { createTRPCReact } from "@trpc/react-query";
 
-export const trpc = createTRPCReact<AppRouter>({
-  abortOnUnmount: true,
-});
+export const trpc = createTRPCReact<AppRouter>({ abortOnUnmount: true });
