+++ apps/web/src/lib/edgestore.ts
@@ -1,7 +1,7 @@
 "use client";
 
-import { type EdgeStoreRouter } from "@/app/api/edgestore/[...edgestore]/route";
 import { createEdgeStoreProvider } from "@edgestore/react";
+import { type EdgeStoreRouter } from "@taskaider/api";
 
 const { EdgeStoreProvider, useEdgeStore } =
   createEdgeStoreProvider<EdgeStoreRouter>();
